Tree-Doctor/api/requirements.txt
tensorflow-macos
tensorflow-metal
fastapi
uvicorn
pillow
numpy
python-multipart
